+++ config/app.config
@@ -4,9 +4,8 @@
 
 monitored_directory: tests/resources/
 
-# Used when running with localq runner to determine the maximum number of
-# used for running checksums
-number_of_cores: 4
+# Determine how many past and active jobs should be kept in memory
+history_len: 100
 
 # Path to the md5sum logs
 md5_log_directory: /tmp/
